CounterMarking

Documentation for CounterMarking.

CounterMarking.meanshow

CounterMarking.randshow

CounterMarking.segment_image

CounterMarking.spots

CounterMarking.stimulus_index

CounterMarking.upperleft

CounterMarking.meanshow — Function

meanshow(seg; kwargs...)

Display a segmented image using the mean color of each segment.

Note

You must load the ImageView package to use this function.

CounterMarking.randshow — Function

randshow(seg; kwargs...)

Display a segmented image using random colors for each segment.

Note

You must load the ImageView package to use this function.

CounterMarking.segment_image — Method

seg = segment_image(img; threshold=0.1, min_size=20)

Given an image img, segment it into regions using a region growing algorithm. min_size is the minimum number of pixels per segment, and threshold determines how different two colors must be to be considered different segments. Larger threshold values will result in fewer segments.

CounterMarking.spots — Method

spotdict, stimulus = spots(seg; max_size_frac=0.1)

Given a segmented image seg, return a Dict(idx => spot) where idx is the segment index and spot is a Spot object where spot.npixels is the number of pixels in the segment and spot.centroid is the centroid of the segment.

stimulus is a Pair{Int, Spot} where the first element is the index of the stimulus segment and the second element is the Spot object for that segment.

Spots larger than max_size_frac * npixels (default: 10% of the image) are ignored.

CounterMarking.stimulus_index — Function

idx = stimulus_index(seg::SegmentedImage, colorproj = RGB(1, 1, -2))

Given a segmented image seg, return the index of the segment that scores highest on the product of (1) projection (dot product) with colorproj and (2) number of pixels.

CounterMarking.upperleft — Method

spotdict, stimulus = upperleft(spotdict::AbstractDict{Int, Spot}, stimulus, imgsize)

Given a spotdict of Spot objects and a stimulus segment, return a new spotdict where the centroids of the spots are flipped so that the stimlus spot is in the upper left corner.